+++ datescontrol/settings.py
@@ -36,6 +36,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'mainapp',
+    'workersapp',
+    'itemtypesapp',
 ]
 
 MIDDLEWARE = [
