+++ mainapp/templates/mainapp/items.html
@@ -6,7 +6,7 @@
         <!--Content top-menu-->
         <ul class="nav shadow-sm">
             <div class="d-flex mt-1 ms-1 mb-1">
-                {% include 'mainapp/includes/brig_checker.html' %}
+                {% include 'includes/brig_checker.html' %}
                 <li class="nav-item dropdown">
                     <select class="form-select-sm">
                         <option selected>Тип</option>
