+++ templates/includes/brig_checker.html
