+++ workersapp/urls.py
@@ -13,18 +13,12 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
-from django.urls import path, include
-from django.conf import settings
-from django.conf.urls.static import static
+from django.urls import path
+
+import workersapp.views as workersapp
+
+app_name = 'workersapp'
 
 urlpatterns = [
-    path('admin/', admin.site.urls, name='admin'),
-
-    path('', include('mainapp.urls', namespace='main')),
-    path('itemtypes/', include('itemtypesapp.urls', namespace='itemtypes')),
-    path('workers/', include('workersapp.urls', namespace='workers')),
+    path('', workersapp.workers, name='index'),
 ]
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
