+++ workersapp/views.py
@@ -2,33 +2,24 @@ from datetime import date, timedelta, datetime
 
 from django.shortcuts import render
 
-from mainapp.models import ItemModel
-from itemtypesapp.models import ItemTypeModel
 from workersapp.models import CityModel, CompanyModel, DepartamentModel, WorkerModel
 
 
 # Create your views here.
-def index(request):
-    context = {'page_title': 'стартовая'}
-    return render(request, 'mainapp/index.html', context)
-
-
-def items(request):
+def workers(request):
     today = datetime.now() + timedelta(hours=5)
     today = date(today.year, today.month, today.day)
-    gived_items = ItemModel.objects.filter(is_active=True, worker__is_active=True, item_type__is_active=True)
+    worker_list = WorkerModel.objects.filter(is_active=True)
     city_list = CityModel.objects.filter(is_active=True)
     company_list = CompanyModel.objects.filter(is_active=True)
     departament_list = DepartamentModel.objects.filter(is_active=True)
     brigadier_list = WorkerModel.objects.filter(is_brig=True, is_active=True)
-    item_list = ItemTypeModel.objects.filter(is_active=True)
-    context = {'page_title': 'список предметов',
-               'items': gived_items,
+    context = {'page_title': 'список сотрудников',
                'cities': city_list,
                'companies': company_list,
                'departaments': departament_list,
                'brigadiers': brigadier_list,
-               'item_types': item_list,
+               'workers': worker_list,
                'today': today,
                }
-    return render(request, 'mainapp/items.html', context)
+    return render(request, 'workersapp/workers.html', context)
